backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pytest
python-dotenv
supabase